ui/termstatus: extract background handling code
