+++ internal/terminal/background_unix_test.go
@@ -1,6 +1,6 @@
 //go:build unix
 
-package termstatus
+package terminal
 
 import (
 	"os"
