+++ internal/terminal/getpgrp_unix.go
@@ -1,6 +1,6 @@
 //go:build unix && !solaris
 
-package termstatus
+package terminal
 
 import "golang.org/x/sys/unix"
