+++ internal/terminal/tcgetpgrp_unix.go
@@ -1,6 +1,6 @@
 //go:build unix && !linux
 
-package termstatus
+package terminal
 
 import "golang.org/x/sys/unix"
