+++ internal/terminal/tcsetpgrp_unix.go
@@ -1,6 +1,6 @@
 //go:build unix && !aix
 
-package termstatus
+package terminal
 
 import "golang.org/x/sys/unix"
